Refactor component imports and remove unused CSS styles
