+++ src/components/capacityDisplay/style.css
@@ -1,0 +1,15 @@
+.capacityCard {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 85%;
+  border: var(--dnb_seaGreen) solid 1px;
+  background-color: var(--dnb_mintGreen);
+  border-radius: 5px;
+  padding: 1rem;
+  margin-bottom: 1rem;
+  margin-top: 0.5rem;
+  color: black;
+  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.497);
+}
